requirements.txt
contourpy==1.2.1
cycler==0.12.1
filelock==3.14.0
fonttools==4.53.0
fsspec==2024.6.0
gate-api==4.26.0
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pykan==0.0.5
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sympy==1.12.1
threadpoolctl==3.5.0
torch==2.2.2
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
zipp==3.19.2
